♻️ noticeList 페이지 useSwr 사용
