+++ package.json
@@ -25,6 +25,7 @@
     "postcss": "8.4.24",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "swr": "^2.2.0",
     "tailwindcss": "3.3.2",
     "typescript": "5.1.6"
   },
